(refactor) move svelte2tsx into subfolder + split up
