+++ packages/svelte2tsx/src/svelte2tsx/knownevents.ts
