+++ packages/svelte2tsx/src/svelte2tsx/nodes/ImplicitTopLevelNames.ts
