+++ packages/svelte2tsx/src/svelte2tsx/nodes/handleScopeAndResolveForSlot.ts
@@ -1,8 +1,8 @@
 import { Node } from 'estree-walker';
-import { BaseDirective, SvelteIdentifier } from '../interfaces';
+import { BaseDirective, SvelteIdentifier } from '../../interfaces';
 import TemplateScope from './TemplateScope';
 import { SlotHandler } from './slot';
-import { isIdentifier, isDestructuringPatterns } from '../utils/svelteAst';
+import { isIdentifier, isDestructuringPatterns } from '../../utils/svelteAst';
 import { extract_identifiers as extractIdentifiers } from 'periscopic';
 
 export function handleScopeAndResolveForSlot({
