+++ packages/svelte2tsx/src/svelte2tsx/utils/tsAst.ts
