+++ packages/svelte2tsx/src/utils/htmlxparser.ts
